feat(mouse): close Residual A geometry for seek-bar and tab-bar (tasks 2.7-2.9)
PlaybackRequest::SeekTo now carries a resolved 0.0..=1.0 fraction that
PlaybackComponent computes against its own painted seekbar_area; the shell
handler App::seek_to_fraction no longer reads layout.playback.seekbar_area.

ensure_tab_visible (a keyboard path) stops reading the painted
layout.tabs_area: the tab-strip text width is now a shared arrangement
primitive (tab_strip_text_width) called by both the chrome painter and
ensure_tab_visible via compute_chrome_geometry.

Adds eligibility-follows-paint and cross-focus seek delivery tests, plus a
doc-comment on fold_mouse_messages recording the single-non-mouse-subscription
invariant (ADR 0024).

+++ src/app/components/msg/playback.rs
@@ -11,7 +11,10 @@ pub enum PlaybackRequest {
     Previous,
     Next,
     SeekRelative(i64),
-    SeekTo(u16),
+    /// Seek to a resolved 0.0..=1.0 fraction of the runtime. `PlaybackComponent`
+    /// resolves the click column against its own `seekbar_area` so no shell code
+    /// reads painted seek-bar geometry (ADR 0022 Residual A).
+    SeekTo(f64),
     ToggleMute,
     VolumeDelta(i64),
     CycleAudio,
